pyproject.toml
[tool.poetry]
name = "aioci"
version = "0.1.0"
description = "Pyaci fork, with asyncio and PEP8 in mind."
authors = ["Jonas Krüger Svensson <jonas.svensson@intility.no>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.22.0"
Flask = "^2.1.2"
lxml = "^4.8.0"
paramiko = "^2.10.4"
parse = "^1.19.0"
pyOpenSSL = "^22.0.0"
PyYAML = "^6.0"
requests = "^2.27.1"
scp = "^0.14.4"
websocket-client = "^1.3.2"
xmltodict = "^0.13.0"
nose = {git = "https://github.com/iress/nose.git"}

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"
black = "^22.3.0"
ipython = "^8.3.0"
coverage = "^6.3.2"
httpretty = "^1.1.4"
sure = "^2.0.0"
pytest = "^7.1.2"

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
    (\.eggs|\.git|\.hg|\.mypy_cache|\.tox|\.venv|\venv|\.github|\docs|\tests|\__pycache__)
)
'''

[tool.isort]
profile = "black"
src_paths = ["fastapi_azure_auth"]
combine_as_imports = true
line_length = 120
sections = [
    'FUTURE',
    'STDLIB',
    'THIRDPARTY',
    'FIRSTPARTY',
    'LOCALFOLDER',
]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
